#76 refactor: date 관련 로직 커스텀 훅으로 분리
